app/pyproject.toml
[tool.poetry]
name = "princetonpy"
version = "0.1.0"
description = "A website for the Princeton Python users group."
authors = ["mhadam <michael@hadam.us>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
django-ses = {extras = ["bounce"], version = "^1.0.3"}
django-newsletter = "^0.9.1"
django-tinymce = "^3.2.0"
sorl-thumbnail = "^12.7.0"
Hypercorn = "^0.11.2"
django-environ = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
